refactor(emitter): split §19 over-cap exports/mod.rs into exported_value_declarations shard
`exports/mod.rs` reached 2194 lines after the DAA fix (+3 lines). Extract
`emit_exported_enum`, `get_enum_member_name`, `emit_enum_value`, and
`emit_exported_variable` (lines 1885–2193, ~310 lines) into a new
under-cap sibling `exported_value_declarations.rs` (319 lines).

Result: `exports/mod.rs` 1889 lines, new shard 319 lines — both under the
§19 2000-line hard cap.
